fix(#63): split ExtractFixturesForLevel into light_fixture_level.cpp
ProposeFixtures (pure clustering) stays in light_fixture_extract.cpp and
unit-testable; the renderer-stack-dependent level wrapper moves to
light_fixture_level.cpp (GLOB-covered in igi-editor, not needed by igi_tests).
